requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
ultralytics
opencv-python
pyperclip
